chore: association overrides for subdecision changes from GEWISDB
See GEWIS/gewisdb#506.
