+++ module/Decision/src/Model/SubDecision.php
@@ -158,20 +158,6 @@ abstract class SubDecision
     private string $contentEN;
 
     /**
-     * The member involved in this sub-decision.
-     *
-     * Not all sub-decisions require this, as such it is nullable. However, sub-decisions that need the guarantee that
-     * this is not null or need to specify an inverse side can do so using an association override.
-     */
-    #[ManyToOne(targetEntity: Member::class)]
-    #[JoinColumn(
-        name: 'lidnr',
-        referencedColumnName: 'lidnr',
-        nullable: true,
-    )]
-    private ?Member $member = null;
-
-    /**
      * Get the decision.
      */
     public function getDecision(): Decision
@@ -238,22 +224,6 @@ abstract class SubDecision
     public function setSequence(int $sequence): void
     {
         $this->sequence = $sequence;
-    }
-
-    /**
-     * Get the member.
-     */
-    public function getMember(): ?Member
-    {
-        return $this->member;
-    }
-
-    /**
-     * Set the member.
-     */
-    public function setMember(Member $member): void
-    {
-        $this->member = $member;
     }
 
     /**
