+++ module/Decision/src/Model/SubDecision/Key/Granting.php
@@ -6,7 +6,9 @@ namespace Decision\Model\SubDecision\Key;
 
 use DateTime;
 use Decision\Model\Keyholder;
+use Decision\Model\Member;
 use Decision\Model\SubDecision;
+use Decision\Model\Trait\MemberAwareTrait;
 use Doctrine\ORM\Mapping\Column;
 use Doctrine\ORM\Mapping\Entity;
 use Doctrine\ORM\Mapping\OneToOne;
@@ -14,6 +16,8 @@ use Doctrine\ORM\Mapping\OneToOne;
 #[Entity]
 class Granting extends SubDecision
 {
+    use MemberAwareTrait;
+
     /**
      * Till when the keycode is granted.
      */
@@ -37,6 +41,16 @@ class Granting extends SubDecision
         mappedBy: 'grantingDec',
     )]
     private Keyholder $keyholder;
+
+    /**
+     * Get the member.
+     *
+     * @psalm-suppress InvalidNullableReturnType
+     */
+    public function getMember(): Member
+    {
+        return $this->member;
+    }
 
     /**
      * Get the date.
