+++ module/Decision/src/Model/Trait/MemberAwareTrait.php
@@ -1,0 +1,42 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Decision\Model\Trait;
+
+use Decision\Model\Member;
+use Doctrine\ORM\Mapping\JoinColumn;
+use Doctrine\ORM\Mapping\ManyToOne;
+
+trait MemberAwareTrait
+{
+    /**
+     * The member involved in this sub-decision.
+     *
+     * Not all sub-decisions require this, as such it is nullable. However, sub-decisions that need the guarantee that
+     * this is not null or need to specify an inverse side can do so using an association override.
+     */
+    #[ManyToOne(targetEntity: Member::class)]
+    #[JoinColumn(
+        name: 'lidnr',
+        referencedColumnName: 'lidnr',
+        nullable: true,
+    )]
+    private ?Member $member = null;
+
+    /**
+     * Get the member.
+     */
+    public function getMember(): ?Member
+    {
+        return $this->member;
+    }
+
+    /**
+     * Set the member.
+     */
+    public function setMember(Member $member): void
+    {
+        $this->member = $member;
+    }
+}
